Feat - created observability modules and moved models to it
